Auto sizing the metadata inputs
Well, the inputs of the metadata div were too large so now they take the width of the content (value or placeholder).

The fact is that all characters aren't at the same width to I made a dummy span where I write the content to know the exact size.

P.S. : And Yeah, I don't want to use Sans-Shérif font.