tried to delete unwanted content
By unwanted content I mean the self promotion of the Website you are taking the light novel (bothersome when after EACH paragraph).

I should let the user choose to enable this functionnality or not because there's a chance that this commit will remove paragraphs of the actual chapter... Please write an issue if you want it. Otherwise it will stay like that because I dont read light novel where the word "light nov" should appear (even then I tried to limit the problems by not putting only "light nov" in the check part).